Remove separate package to make installation simpler
