+++ src/lib/gui.py
@@ -14,9 +14,8 @@ from typing import Callable, Literal, Union
 from dataclasses import dataclass
 
 #import sys and use path insert to add lib files
-from .img_lib import image_to_tk_image, fit_image, RGBA_to_RGB, LA_to_L
-from .backend_lib import Smart_Image
-from .tuple_utils import *
+from .img import image_to_tk_image, fit_image, RGBA_to_RGB, LA_to_L
+from .tuple import *
 #endregion
 
 # widget to display info, errors, warning, and text
@@ -95,156 +94,6 @@ class Debug():
     self.__widget__.config(fg = colors[0],
                        bg = colors[1])
 
-@dataclass
-class CycleState:
-  text: str
-  fg_bg: tuple[str, str]
-  enter_exit: tuple[Callable | None, Callable | None]
-
-# creates a button that cycles through states
-# each state contains: text, colors, and entering / exiting functions
-class Cycle():
-  # global counter to ensure unique names
-  __total_cycles__: int = 0
-  # mandatory / main fields
-  __widget__: Button
-  __gui__: GUI_Controller
-  state: int
-  # user-inputted fields
-  __states__: list[CycleState]
-  func_always: Callable | None
-  
-  # create new Cycle widget
-  def __init__( self, gui: GUI_Controller,
-          name: str | None = None,
-          func_always: Callable | None = None):
-    self.__gui__ = gui
-    self.func_always = func_always
-    self.state = 0
-    self.__states__ = []
-    self.__widget__ = Button(gui.root, command=self.cycle)
-    if name is None:
-      gui.add_widget(f"unnamed cycle widget {Cycle.__total_cycles__}", self)
-      Cycle.__total_cycles__ += 1
-    else:
-      gui.add_widget(name, self)
-    
-  # place widget on the grid
-  def grid(self, row: int | None = None, col: int | None = None, colspan: int = 1, rowspan: int = 1):
-    if row is None or col is None:
-      self.__widget__.grid()
-    else:
-      self.__widget__.grid(row = row,
-                       column = col,
-                       rowspan = rowspan,
-                       columnspan = colspan,
-                       sticky = "nesw")
-      
-  # remove widget from the grid
-  def grid_remove(self):
-    self.__widget__.grid_remove()
-    
-  # add a new state to the cycle
-  def add_state(self,
-                text: str = "",
-                colors: tuple[str,str] = ("black", "white"),
-                enter: Callable | None = None,
-                exit: Callable | None = None):
-    self.__states__.append(CycleState(text, colors, (enter, exit)))
-    # update widget cosmetics to the first state added
-    if len(self.__states__) == 1:
-      self.__widget__.config( text = text,
-                          fg = colors[0],
-                          bg = colors[1])
-  
-  # get a state in cycle_state_t format with added index:
-  # (index, cycle_state_t) = (index, (text, colors, (enter, exit)))
-  def get_state(self, index: int) -> CycleState | None:
-    try:
-      return self.__states__[index]
-    except IndexError:
-      if self.__gui__.debug is not None:
-        self.__gui__.debug.error(f"invalid cycle get target {index} > {len(self.__states__)-1}")
-      return None
-  
-  # jump to a state index
-  def goto(self, index: int = 0):
-    # check if index is valid
-    if index < 0 or index >= len(self.__states__):
-      if self.__gui__.debug is not None:
-        self.__gui__.debug.error(f"invalid cycle update target {index} > {len(self.__states__)-1}")
-      return
-
-    # call exit function of previous state
-    if self.__states__[self.state].enter_exit[1] is not None:
-      self.__states__[self.state].enter_exit[1]()
-    # update state
-    self.state = index
-    self.__widget__.config( text = self.__states__[index].text,
-                        fg = self.__states__[index].fg_bg[0],
-                        bg = self.__states__[index].fg_bg[1])
-    if self.__gui__.debug is not None:
-      self.__gui__.debug.info(f"Cycle set to {index}: "+self.__states__[index].text.split("\n")[0])
-    # call enter function of new state
-    if self.__states__[index].enter_exit[0] is not None:
-      self.__states__[index].enter_exit[0]()
-    # call always function if specified
-    if self.func_always is not None:
-      self.func_always()
-      
-  # update the cycle to reflext current state
-  # useful when using update_state
-  def update(self):
-    self.__widget__.config(text = self.__states__[self.state].text,
-                           fg = self.__states__[self.state].fg_bg[0],
-                           bg = self.__states__[self.state].fg_bg[1])
-  
-  def cycle(self):
-    if(len(self.__states__) == 0):
-      if(self.__gui__.debug != None):
-        self.__gui__.debug.warn("Tried to cycle with no states")
-      return
-    self.goto((self.state + 1) % len(self.__states__))
-
-  def state_name(self) -> str:
-    return self.__states__[self.state].text
-  
-  def get_state_names(self) -> list[str]:
-    return [state.text for state in self.__states__]
-  
-  def update_state(self, state: int,
-                   text: str | None = None,
-                   colors: tuple[str,str] | None = None,
-                   enter_func: Callable | None = None,
-                   exit_func: Callable | None = None,
-                   state_type: CycleState | None = None):
-    # check if index is valid
-    if state < 0 or state >= len(self.__states__):
-      if(self.__gui__.debug != None):
-        self.__gui__.debug.error(f"invalid cycle update target {state} > {len(self.__states__)-1}")
-      return
-    
-    #if state_type is specified, set to that, update, and return
-    if state_type is not None:
-      self.__states__[state] = state_type
-      self.__widget__.config( text = state_type.text,
-                          fg = state_type.fg_bg[0],
-                          bg = state_type.fg_bg[1])
-    else:
-      # update state
-      temp_state = self.__states__[state]
-      if text is not None:
-        temp_state.text = text
-      if colors is not None:
-        temp_state.fg_bg = colors
-      temp_state.enter_exit = ( enter_func if enter_func is not None else temp_state.enter_exit[0],
-                        exit_func if exit_func is not None else temp_state.enter_exit[1])
-      self.__states__[state] = temp_state
-    
-    # if current state, update text forcefully
-    if state == self.state:
-      self.update()
-      
 # creates thumbnail / image import widget
 class Thumbnail(): 
   widget: Button
@@ -699,7 +548,7 @@ class TextPopup():
 # GUI controller and widget manager
 class GUI_Controller():
   # list of accepted widget types
-  gui_widgets = Union[Widget, Cycle, Thumbnail, Debug, Intput, Floatput, TextPopup]
+  gui_widgets = Union[Widget, Thumbnail, Debug, Intput, Floatput, TextPopup]
   #region: fields
   ### Internal Fields ###
   root: Tk
@@ -818,123 +667,3 @@ class GUI_Controller():
         return result
       else:
         return div(result, widget_size)
-  
-# swaps around groups of widgets
-# Requires a GUI controller
-# hides widgets with grid_remove()
-class Smart_Area():
-  # gui controller to query widgets from
-  __gui__: GUI_Controller
-  # list of widget lists
-  __widgets__: list[list[GUI_Controller.gui_widgets]]
-  # list of special fucntions to call when switching groups
-  # show first, hide second (show, hide)
-  __funcs__: list[tuple[Callable | None, Callable | None]]
-  # current index of the list
-  __index__: int = 0
-  # debug widget
-  debug: Debug | None = None
-  name: str | None = None
-  #endregion
-  
-  def __init__( self,
-                gui: GUI_Controller,
-                debug: Debug | None = None,
-                name: str | None = None):
-    self.__widgets__ = []
-    self.__funcs__ = []
-    self.__gui__ = gui
-    self.debug = debug
-    self.name = name
-    
-  def __hide__(self, group: int, func_first: bool = True):
-    if(func_first and self.__funcs__[group][1] != None):
-      self.__funcs__[group][1]()
-    for widget in self.__widgets__[group]:
-      widget.grid_remove()
-    if(not func_first and self.__funcs__[group][1] != None):
-      self.__funcs__[group][1]()
-    
-  def __show__(self, group: int, func_first: bool = True):
-    if(func_first and self.__funcs__[group][0] != None):
-      self.__funcs__[group][0]()
-    for widget in self.__widgets__[group]:
-      widget.grid()
-    if(not func_first and self.__funcs__[group][0] != None):
-      self.__funcs__[group][0]()
-  
-  # add one or several widgets to specified group
-  # disable hide to leave widgets visible after adding (not recommended, use jump() instead)
-  def add(self, group: int, widgets: str | list[str], hide: bool = True):
-    #normalize to list
-    names: list[str]
-    if(type(widgets) == str):
-      names = [widgets]
-    elif(type(widgets) == list):
-      names = widgets
-    else:
-      raise Exception("widgets must be a string or a list of strings")
-    
-    # check group number
-    if(group < 0):
-      raise Exception("only positive group numbers are allowed")
-    elif(group > len(self.__widgets__)):
-      if(self.debug != None):
-        msg: str = "non-consecutive group number, added filler groups"
-        if(self.name != None):
-          msg = self.name+" "+msg
-        self.debug.warn(msg)
-      for i in range(len(self.__widgets__), group+1):
-        self.__widgets__.append([])
-        self.__funcs__.append((None, None))
-    elif(group == len(self.__widgets__)):
-      self.__widgets__.append([])
-      self.__funcs__.append((None, None))
-      
-    # add widgets to list
-    for name in names:
-      widget: GUI_Controller.gui_widgets | None = self.__gui__.get_widget(name)
-      if(widget == None):
-        if(self.debug != None):
-          msg: str = "Tried to add non-existent widget \""+name+"\" to group "+str(group)
-          if(self.name != None):
-            msg = self.name+" "+msg
-          self.debug.error(msg)
-      else:
-        self.__widgets__[group].append(widget)
-    
-    if(hide):
-      self.__hide__(group)
-  
-  # add special functions when showing or hiding a group
-  def add_func(self, group: int, show: Callable | None = None, hide: Callable | None = None):
-    if(group < 0 or group >= len(self.__funcs__)):
-      if(self.debug != None):
-        msg: str = "can only add special functions to existing groups, add an empty group first"
-        if(self.name != None):
-          msg = self.name+" "+msg
-        self.debug.error(msg)
-      return
-    self.__funcs__[group] = (show, hide)
-    
-  
-  # hide current widget and jump to specified group
-  # if order matters, toggle the func_first parameter (show, hide)
-  def jump(self, group: int, func_first: tuple[bool,bool] = (True, True)):
-    if(self.debug != None):
-      msg: str = "switched from group "+str(self.__index__)+" to "+str(group % len(self.__widgets__))
-      if(self.name != None):
-        msg = self.name+" "+msg
-      self.debug.info(msg)
-    self.__hide__(self.__index__, func_first[1])
-    self.__index__ = group % len(self.__widgets__)
-    self.__show__(self.__index__, func_first[0])
-  
-  def next(self, func_first: tuple[bool,bool] = (True, True)):
-    self.jump(self.__index__ + 1, func_first)
-    
-  def prev(self, func_first: tuple[bool,bool] = (True, True)):
-    self.jump(self.__index__ - 1, func_first)
-
-  def index(self) -> int:
-    return self.__index__
